pyproject.toml
[project]
name = "edu.princeton.tango"
version = "0.0.1"
authors = [
    { name = "Benjamin Herber", email = "bherber@princeton.edu" },
    { name = "Maria Apostolaki", email = "apostolaki@princeton.edu" },
    { name = "Henry Birge-Lee", email = "birgelee@princeton.edu " },
]
description = "An entrypoint to the Tango cooperative peer routing solution"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = ["pyyaml >= 6.0, < 7"]

[project.optional-dependencies]
dev = [
    "pytest >= 7.2.2, < 8",
    "black >= 23.1.0, < 24",
    "isort >= 5.12.0, < 6",
    "tox >= 4.4.6, < 5"
]

[project.urls]
"Homepage" = "https://github.com/PrincetonUniversity/tango-routing"
"Bug Tracker" = "https://github.com/PrincetonUniversity/tango-routing/issues"

[tool.setuptools.packages.find]
where = ["src/python"]

[tool.black]
line-length = 79
exclude = """
/(
    .eggs
  | .git
  | .hg
  | .mypy_cache
  | .tox
  | .venv
  | _build
  | buck-out
  | build
  | dist
)/
"""

[tool.isort]
profile = "black"
combine_as_imports = true
line_length = 79
